refactor: VisualTransformation 패키지 이동
