+++ app/src/main/java/nextstep/payments/screens/card/update/components/util/CardNumberVisualTransformation.kt
@@ -1,25 +1,29 @@
-package nextstep.payments.screens.card.new
+package nextstep.payments.screens.card.update.components.util
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.OffsetMapping
 import androidx.compose.ui.text.input.TransformedText
 import androidx.compose.ui.text.input.VisualTransformation
 
-class ExpiredDateVisualTransformation: VisualTransformation {
+class CardNumberVisualTransformation : VisualTransformation {
     private val offsetMapping = object : OffsetMapping {
         override fun originalToTransformed(offset: Int): Int {
             return when (offset) {
-                in 0..2 -> offset
-                in 3..4 -> offset + 1
-                else -> offset + 1
+                in 0..4 -> offset
+                in 5..8 -> offset + 1
+                in 9..12 -> offset + 2
+                in 13..15 -> offset + 3
+                else -> offset + 3
             }
         }
 
         override fun transformedToOriginal(offset: Int): Int {
             return when (offset) {
-                in 0..2 -> offset
-                in 3..4 -> offset - 1
-                else -> offset - 1
+                in 0..3 -> offset
+                in 5..8 -> offset - 1
+                in 10..13 -> offset - 2
+                in 15..18 -> offset - 3
+                else -> offset - 3
             }
         }
     }
@@ -38,7 +42,7 @@ class ExpiredDateVisualTransformation: VisualTransformation {
     }
 
     companion object {
-        private const val GROUP_SIZE = 2
-        private const val SEPARATOR: String = "/"
+        private const val GROUP_SIZE = 4
+        private const val SEPARATOR: String = "-"
     }
 }
