+++ app/src/main/java/nextstep/payments/screens/card/update/components/util/ExpiredDateVisualTransformation.kt
@@ -1,29 +1,25 @@
-package nextstep.payments.screens.card.new
+package nextstep.payments.screens.card.update.components.util
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.OffsetMapping
 import androidx.compose.ui.text.input.TransformedText
 import androidx.compose.ui.text.input.VisualTransformation
 
-class CardNumberVisualTransformation : VisualTransformation {
+class ExpiredDateVisualTransformation: VisualTransformation {
     private val offsetMapping = object : OffsetMapping {
         override fun originalToTransformed(offset: Int): Int {
             return when (offset) {
-                in 0..4 -> offset
-                in 5..8 -> offset + 1
-                in 9..12 -> offset + 2
-                in 13..15 -> offset + 3
-                else -> offset + 3
+                in 0..2 -> offset
+                in 3..4 -> offset + 1
+                else -> offset + 1
             }
         }
 
         override fun transformedToOriginal(offset: Int): Int {
             return when (offset) {
-                in 0..3 -> offset
-                in 5..8 -> offset - 1
-                in 10..13 -> offset - 2
-                in 15..18 -> offset - 3
-                else -> offset - 3
+                in 0..2 -> offset
+                in 3..4 -> offset - 1
+                else -> offset - 1
             }
         }
     }
@@ -42,7 +38,7 @@ class CardNumberVisualTransformation : VisualTransformation {
     }
 
     companion object {
-        private const val GROUP_SIZE = 4
-        private const val SEPARATOR: String = "-"
+        private const val GROUP_SIZE = 2
+        private const val SEPARATOR: String = "/"
     }
 }
